Update README.md
In a galaxy of power struggles, the stealthy Elgem race investigates a Khaured threat for the Galactic Federation. Tareth leads a team of Loria, Velryn, and Selene to a Khaured outpost, uncovering a powerful warship with bio-organic armor. They steal its schematics and escape despite detection. Back in Elgem space, analysis reveals the ship’s link to ancient Oshi technology, prompting Tareth to alert the Federation and prepare for independent action as the galaxy faces an escalating threat.